Refactor thread export into guarded package
