+++ src/mindroom/thread_export/__init__.py
@@ -1,0 +1,11 @@
+"""Export Matrix threads to durable, searchable workspace files."""
+
+from mindroom.thread_export.models import ThreadExportStats, ThreadExportTarget
+from mindroom.thread_export.service import export_threads_once, export_threads_to_targets_once
+
+__all__ = [
+    "ThreadExportStats",
+    "ThreadExportTarget",
+    "export_threads_once",
+    "export_threads_to_targets_once",
+]
